requirements.txt
encodings
os
importlib
collections
tkinter
idlelib
xml
pyexpat
urllib
numpy
json
ctypes
opencv-python
scikit-learn
scipy
typing
joblib
multiprocessing
concurrent
psutil
unittest
html
tensorflow
absl-py
google
wrapt
gast
six
astunparse
termcolor
opt-einsum
packaging
email
http
urllib3
charset-normalizer
requests
idna
certifi
h5py
asyncio
distutils
flatbuffers
tensorboard
fsspec
keras
Pillow
cffi
pytz
python-dateutil
pandas
pyarrow
nltk
regex
sqlite3
thinc
pydantic
srsly
ruamel.yaml
confection
torch
tqdm
mpmath
sympy
wasabi
colorama
blis
cymem
preshed
murmurhash
spacy
langcodes
click
typer
shellingham
rich
attr
markdown-it-py
mdurl
uc-micro
linkify-it-py
pygments
jinja2
markupsafe
annoy
rubicon-objc
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-Foundation
pyobjc-framework-AppKit
pyobjc-framework-Quartz
pyautogui
transformers
huggingface-hub
PyYAML
filelock
sentencepiece
tokenizers
SafeTensors
pytesseract
screeninfo